cloud/requirements.txt
fitbit==0.3.1
azure-iot-hub==2.6.1
python-dateutil==2.8.2
python-dotenv
numpy
